Refactor pegen.py into multiple files
To organize the project, split pegen.py into individual components
and add a package structure with a main entrypoint.
