+++ pegen/grammar.py
@@ -1,16 +1,6 @@
-#!/usr/bin/env python3.8
-
-"""pegen -- PEG Generator.
-
-Search the web for PEG Parsers for reference.
-"""
-
 from __future__ import annotations  # Requires Python 3.7 or later
 
-import argparse
 import ast
-import contextlib
-import os
 import re
 import sys
 import time
@@ -19,333 +9,18 @@ import tokenize
 import traceback
 from typing import AbstractSet, Callable, Dict, Generic, Iterable, Optional, List, TypeVar, Union
 
-import sccutils
-
-T = TypeVar('T')
-
-exact_token_types = token.EXACT_TOKEN_TYPES  # type: ignore
-
-Mark = int  # NewType('Mark', int)
+from pegen.parser import memoize, Parser
+from pegen.tokenizer import exact_token_types
 
 
-def shorttok(tok: tokenizer.TokenInfo) -> str:
-    return "%-25.25s" % f"{tok.start[0]}.{tok.start[1]}: {token.tok_name[tok.type]}:{tok.string!r}"
-
-
-class Tokenizer:
-    """Caching wrapper for the tokenize module.
-
-    This is pretty tied to Python's syntax.
-    """
-
-    _tokens: List[tokenize.TokenInfo]
-
-    def __init__(self, tokengen: Iterable[TokenInfo], *, verbose=False):
-        self._tokengen = tokengen
-        self._tokens = []
-        self._index = 0
-        self._verbose = verbose
-        if verbose:
-            self.report(False, False)
-
-    def getnext(self) -> tokenize.TokenInfo:
-        """Return the next token and updates the index."""
-        cached = True
-        while self._index == len(self._tokens):
-            tok = next(self._tokengen)
-            if tok.type in (token.NL, token.COMMENT):
-                continue
-            if tok.type == token.ERRORTOKEN and tok.string.isspace():
-                continue
-            self._tokens.append(tok)
-            cached = False
-        tok = self._tokens[self._index]
-        self._index += 1
-        if self._verbose:
-            self.report(cached, False)
-        return tok
-
-    def peek(self) -> tokenize.TokenInfo:
-        """Return the next token *without* updating the index."""
-        while self._index == len(self._tokens):
-            tok = next(self._tokengen)
-            if tok.type in (token.NL, token.COMMENT):
-                continue
-            if tok.type == token.ERRORTOKEN and tok.string.isspace():
-                continue
-            self._tokens.append(tok)
-        return self._tokens[self._index]
-
-    def diagnose(self) -> tokenize.TokenInfo:
-        if not self._tokens:
-            self.getnext()
-        return self._tokens[-1]
-
-    def mark(self) -> Mark:
-        return self._index
-
-    def reset(self, index: Mark) -> None:
-        if index == self._index:
-            return
-        assert 0 <= index <= len(self._tokens), (index, len(self._tokens))
-        old_index = self._index
-        self._index = index
-        if self._verbose:
-            self.report(True, index < old_index)
-
-    def report(self, cached, back):
-        if back:
-            fill = '-'*self._index + '-'
-        elif cached:
-            fill = '-'*self._index + '>'
-        else:
-            fill = '-'*self._index + '*'
-        if self._index == 0:
-            print(f"{fill} (Bof)")
-        else:
-            tok = self._tokens[self._index - 1]
-            print(f"{fill} {shorttok(tok)}")
-
-
-def memoize(method: Callable[[Parser], T]):
-    """Memoize a symbol method."""
-    method_name = method.__name__
-
-    def symbol_wrapper(self: Parser) -> T:
-        mark = self.mark()
-        key = mark, method_name
-        # Fast path: cache hit, and not verbose.
-        if key in self._symbol_cache and not self._verbose:
-            tree, endmark = self._symbol_cache[key]
-            if tree:
-                self.reset(endmark)
-            else:
-                assert mark == endmark
-            return tree
-        # Slow path: no cache hit, or verbose.
-        verbose = self._verbose
-        fill = '  ' * self._level
-        if key not in self._symbol_cache:
-            if verbose:
-                print(f"{fill}{method_name} ... (looking at {self.showpeek()})")
-            self._level += 1
-            tree = method(self)
-            self._level -= 1
-            if verbose:
-                print(f"{fill}... {method_name} -> {tree!s:.100}")
-            if tree:
-                endmark = self.mark()
-            else:
-                endmark = mark
-                self.reset(endmark)
-            self._symbol_cache[key] = tree, endmark
-        else:
-            tree, endmark = self._symbol_cache[key]
-            if verbose:
-                print(f"{fill}{method_name} -> {tree!s:.100}")
-            if tree:
-                self.reset(endmark)
-        return tree
-
-    symbol_wrapper.__wrapped__ = method
-    return symbol_wrapper
-
-
-def memoize_left_rec(method: Callable[[Parser], T]):
-    """Memoize a left-recursive symbol method."""
-    method_name = method.__name__
-
-    def symbol_wrapper(self: Parser) -> T:
-        mark = self.mark()
-        key = mark, method_name
-        # Fast path: cache hit, and not verbose.
-        if key in self._symbol_cache and not self._verbose:
-            tree, endmark = self._symbol_cache[key]
-            if tree:
-                self.reset(endmark)
-            else:
-                assert mark == endmark
-            return tree
-        # Slow path: no cache hit, or verbose.
-        verbose = self._verbose
-        fill = '  ' * self._level
-        if key not in self._symbol_cache:
-            if verbose:
-                print(f"{fill}{method_name} ... (looking at {self.showpeek()})")
-            self._level += 1
-
-            # For left-recursive rules we manipulate the cache and
-            # loop until the rule shows no progress, then pick the
-            # previous result.  For an explanation why this works, see
-            # https://github.com/PhilippeSigaud/Pegged/wiki/Left-Recursion
-            # (But we use the memoization cache instead of a static
-            # variable.)
-
-            # Prime the cache with a failure.
-            self._symbol_cache[key] = None, mark
-            lastresult, lastmark = None, mark
-            depth = 0
-            if verbose:
-                print(f"{fill}Recursive {method_name} at {mark} depth {depth}")
-
-            while True:
-                self.reset(mark)
-                result = method(self)
-                endmark = self.mark()
-                depth += 1
-                if verbose:
-                    print(f"{fill}Recursive {method_name} at {mark} depth {depth}: {result!s:.100} to {endmark}")
-                if not result:
-                    if verbose:
-                        print(f"{fill}Fail with {lastresult!s:.100} to {lastmark}")
-                    break
-                if endmark <= lastmark:
-                    if verbose:
-                        print(f"{fill}Bailing with {lastresult!s:.100} to {lastmark}")
-                    break
-                self._symbol_cache[key] = lastresult, lastmark = result, endmark
-
-            self.reset(lastmark)
-            tree = lastresult
-
-            self._level -= 1
-            if verbose:
-                print(f"{fill}{method_name} -> {tree!s:.100}")
-            if tree:
-                endmark = self.mark()
-            else:
-                endmark = mark
-                self.reset(endmark)
-            self._symbol_cache[key] = tree, endmark
-        else:
-            tree, endmark = self._symbol_cache[key]
-            if verbose:
-                print(f"{fill}{method_name} -> {tree!s:.100}")
-            if tree:
-                self.reset(endmark)
-        return tree
-
-    symbol_wrapper.__wrapped__ = method
-    return symbol_wrapper
-
-
-def memoize_expect(method: Callable[[Parser], Optional[tokenize.TokenInfo]]) -> bool:
-    """Memoize the expect() method."""
-
-    def expect_wrapper(self: Parser, type: str) -> Optional[tokenize.TokenInfo]:
-        mark = self.mark()
-        key = mark, type
-        # Fast path: cache hit.
-        if key in self._token_cache:
-            res, endmark = self._token_cache[key]
-            if res:
-                self.reset(endmark)
-                # Uncomment these when parsing Python, to save
-                # up to 80% of memory (though little time!).
-                # self._token_cache.clear()
-                # self._symbol_cache.clear()
-            return res
-        # Slow path.
-        if key not in self._token_cache:
-            res = method(self, type)
-            if res:
-                endmark = self.mark()
-            else:
-                endmark = mark
-            self._token_cache[key] = res, endmark
-        else:
-            res, endmark = self._token_cache[key]
-        self.reset(endmark)
-        return res
-
-    expect_wrapper.__wrapped__ = method
-    return expect_wrapper
-
-
-class Parser(Generic[T]):
-    """Parsing base class."""
-
-    def __init__(self, tokenizer: Tokenizer, *, verbose=False):
-        self._tokenizer = tokenizer
-        self._verbose = verbose
-        self._level = 0
-        self._symbol_cache: Dict[Tuple[Mark,
-                                       Callable[[Parser], Optional[T]]],
-                                 Tuple[Optional[T], Mark]] = {}
-        self._token_cache: Dict[Tuple[Mark, str], bool] = {}
-        # Pass through common tokeniser methods.
-        # TODO: Rename to _mark and _reset.
-        self.mark = self._tokenizer.mark
-        self.reset = self._tokenizer.reset
-
-    def showpeek(self):
-        tok = self._tokenizer.peek()
-        return f"{tok.start[0]}.{tok.start[1]}: {token.tok_name[tok.type]}:{tok.string!r}"
-
-    def cut(self):
-        if self._verbose:
-            fill = '  ' * self._level
-            print(f"{fill}CUT ... (looking at {self.showpeek()})")
-        return True
-
-    @memoize
-    def name(self) -> Optional[tokenize.TokenInfo]:
-        tok = self._tokenizer.peek()
-        if tok.type == token.NAME:
-            return self._tokenizer.getnext()
-        return None
-
-    @memoize
-    def number(self) -> Optional[tokenize.TokenInfo]:
-        tok = self._tokenizer.peek()
-        if tok.type == token.NUMBER:
-            return self._tokenizer.getnext()
-        return None
-
-    @memoize
-    def string(self) -> Optional[tokenize.TokenInfo]:
-        tok = self._tokenizer.peek()
-        if tok.type == token.STRING:
-            return self._tokenizer.getnext()
-        return None
-
-    @memoize
-    def curly_stuff(self) -> Optional[tokenize.TokenInfo]:
-        tok = self._tokenizer.peek()
-        if tok.type == CURLY_STUFF:
-            return self._tokenizer.getnext()
-        return None
-
-    @memoize_expect
-    def expect(self, type: str) -> Optional[tokenize.TokenInfo]:
-        tok = self._tokenizer.peek()
-        if tok.string == type:
-            return self._tokenizer.getnext()
-        if type in exact_token_types:
-            if tok.type == exact_token_types[type]:
-                return self._tokenizer.getnext()
-        if type in token.__dict__:
-            if tok.type == token.__dict__[type]:
-                return self._tokenizer.getnext()
-        if tok.type == token.OP and tok.string == type:
-            return self._tokenizer.getnext()
-        return None
-
-    def positive_lookahead(self, func: Callable[..., T], *args) -> Optional[T]:
-        mark = self.mark()
-        ok = func(*args)
-        self.reset(mark)
-        return ok
-
-    def negative_lookahead(self, func: Callable[..., T], *args) -> bool:
-        mark = self.mark()
-        ok = func(*args)
-        self.reset(mark)
-        return not ok
-
-    def make_syntax_error(self, filename="<unknown>") -> NoReturn:
-        tok = self._tokenizer.diagnose()
-        return SyntaxError("pegen parse failure", (filename, tok.start[0], 1 + tok.start[1], tok.line))
+def dedupe(name: str, names: List[str]) -> str:
+    origname = name
+    counter = 0
+    while name in names:
+        counter += 1
+        name = f"{origname}_{counter}"
+    names.append(name)
+    return name
 
 
 class Rule:
@@ -1157,439 +832,3 @@ class GrammarParser(Parser):
             return StringLeaf(string.string)
         return None
 
-
-MODULE_PREFIX = """\
-#!/usr/bin/env python3.8
-# @generated by pegen.py from {filename}
-from __future__ import annotations
-
-import ast
-import sys
-import tokenize
-
-from pegen import memoize, memoize_left_rec, Parser
-
-"""
-
-MODULE_SUFFIX = """
-
-if __name__ == '__main__':
-    from pegen import simple_parser_main
-    simple_parser_main(GeneratedParser)
-"""
-
-EXTENSION_PREFIX = """\
-// @generated by pegen.py from {filename}
-
-#include "pegen.h"
-"""
-
-EXTENSION_SUFFIX = """
-// TODO: Allow specifying a module name
-
-static PyObject *
-parse_file(PyObject *self, PyObject *args)
-{
-    const char *filename;
-
-    if (!PyArg_ParseTuple(args, "s", &filename))
-        return NULL;
-    return run_parser(filename, (void *)start_rule, %(mode)s);
-}
-
-static PyMethodDef ParseMethods[] = {
-    {"parse",  parse_file, METH_VARARGS, "Parse a file."},
-    {NULL, NULL, 0, NULL}        /* Sentinel */
-};
-
-static struct PyModuleDef parsemodule = {
-    PyModuleDef_HEAD_INIT,
-    .m_name = "parse",
-    .m_doc = "A parser.",
-    .m_methods = ParseMethods,
-};
-
-PyMODINIT_FUNC
-PyInit_parse(void)
-{
-    PyObject *m = PyModule_Create(&parsemodule);
-    if (m == NULL)
-        return NULL;
-
-    return m;
-}
-
-// The end
-"""
-
-
-class ParserGenerator:
-
-    def __init__(self, rules: Dict[str, Rule], file: Optional[IO[Text]]):
-        self.rules = rules
-        self.file = file
-        self.level = 0
-        compute_nullables(rules)
-        self.first_graph, self.first_sccs = compute_left_recursives(self.rules)
-        self.todo = self.rules.copy()  # Rules to generate
-        self.counter = 0  # For name_rule()/name_loop()
-
-    @contextlib.contextmanager
-    def indent(self) -> None:
-        self.level += 1
-        try:
-            yield
-        finally:
-            self.level -= 1
-
-    def print(self, *args):
-        if not args:
-            print(file=self.file)
-        else:
-            print("    "*self.level, end="", file=self.file)
-            print(*args, file=self.file)
-
-    def printblock(self, lines):
-        for line in lines.splitlines():
-            self.print(line)
-
-    def generate_python_module(self, filename: str) -> None:
-        self.print(MODULE_PREFIX.format(filename=filename))
-        self.print("class GeneratedParser(Parser):")
-        while self.todo:
-            for rulename, rule in list(self.todo.items()):
-                del self.todo[rulename]
-                self.print()
-                with self.indent():
-                    rule.pgen_func(self)
-        self.print(MODULE_SUFFIX.rstrip('\n'))
-
-    def generate_cpython_extension(self, filename: str) -> None:
-        self.collect_todo()
-        self.print(EXTENSION_PREFIX.format(filename=filename))
-        for i, rulename in enumerate(self.todo, 1000):
-            self.print(f"#define {rulename}_type {i}")
-        self.print()
-        for rulename, rule in self.todo.items():
-            if rule.is_loop():
-                type = 'asdl_seq *'
-            elif rule.type:
-                type = rule.type + ' '
-            else:
-                type = 'void *'
-            self.print(f"static {type}{rulename}_rule(Parser *p);")
-        self.print()
-        while self.todo:
-            for rulename, rule in list(self.todo.items()):
-                del self.todo[rulename]
-                self.print()
-                rule.cgen_func(self)
-        mode = int(self.rules['start'].type == 'mod_ty')
-        self.print(EXTENSION_SUFFIX.rstrip('\n') % dict(mode=mode))
-
-    def collect_todo(self) -> None:
-        done = set()  # type: Set[str]
-        while True:
-            alltodo = set(self.todo)
-            todo = alltodo - done
-            if not todo:
-                break
-            for rulename in todo:
-                self.todo[rulename].collect_todo(self)
-            done = alltodo
-
-    def name_node(self, rhs: Rhs) -> str:
-        self.counter += 1
-        name = f'_tmp_{self.counter}'  # TODO: Pick a nicer name.
-        self.todo[name] = Rule(name, None, rhs)
-        return name
-
-    def name_loop(self, node: Plain, is_repeat1: bool) -> str:
-        self.counter += 1
-        if is_repeat1:
-            prefix = '_loop1_'
-        else:
-            prefix = '_loop0_'
-        name = f'{prefix}{self.counter}'  # TODO: It's ugly to signal via the name.
-        self.todo[name] = Rule(name, None, Rhs([Alt([NamedItem(None, node)])]))
-        return name
-
-
-def compute_nullables(rules: Dict[str, Rule]) -> None:
-    """Compute which rules in a grammar are nullable.
-
-    Thanks to TatSu (tatsu/leftrec.py) for inspiration.
-    """
-    for rule in rules.values():
-        rule.visit(rules)
-
-
-def compute_left_recursives(rules: Dict[str, Rule]) -> Tuple[Dict[str, Set[str]], List[Set[str]]]:
-    graph = make_first_graph(rules)
-    sccs = list(sccutils.strongly_connected_components(graph.keys(), graph))
-    for scc in sccs:
-        if len(scc) > 1:
-            for name in scc:
-                rules[name].left_recursive = True
-            # Try to find a leader such that all cycles go through it.
-            leaders = set(scc)
-            for start in scc:
-                for cycle in sccutils.find_cycles_in_scc(graph, scc, start):
-                    ## print("Cycle:", " -> ".join(cycle))
-                    leaders -= (scc - set(cycle))
-                    if not leaders:
-                        raise ValueError(f"SCC {scc} has no leadership candidate (no element is included in all cycles)")
-            ## print("Leaders:", leaders)
-            leader = min(leaders)  # Pick an arbitrary leader from the candidates.
-            rules[leader].leader = True
-        else:
-            name = min(scc)  # The only element.
-            if name in graph[name]:
-                rules[name].left_recursive = True
-                rules[name].leader = True
-    return graph, sccs
-
-
-def make_first_graph(rules: Dict[str, Rule]) -> Dict[str, str]:
-    """Compute the graph of left-invocations.
-
-    There's an edge from A to B if A may invoke B at its initial
-    position.
-
-    Note that this requires the nullable flags to have been computed.
-    """
-    graph = {}
-    vertices = set()
-    for rulename, rhs in rules.items():
-        graph[rulename] = names = rhs.initial_names()
-        vertices |= names
-    for vertex in vertices:
-        graph.setdefault(vertex, set())
-    return graph
-
-
-def dedupe(name: str, names: List[str]) -> str:
-    origname = name
-    counter = 0
-    while name in names:
-        counter += 1
-        name = f"{origname}_{counter}"
-    names.append(name)
-    return name
-
-
-def print_memstats() -> bool:
-    MiB: Final = 2**20
-    try:
-        import psutil
-    except ImportError:
-        return False
-    print("Memory stats:")
-    process = psutil.Process()
-    meminfo = process.memory_info()
-    res = {}
-    res['rss'] = meminfo.rss / MiB
-    res['vms'] = meminfo.vms / MiB
-    if sys.platform == 'win32':
-        res['maxrss'] = meminfo.peak_wset / MiB
-    else:
-        # See https://stackoverflow.com/questions/938733/total-memory-used-by-python-process
-        import resource  # Since it doesn't exist on Windows.
-        rusage = resource.getrusage(resource.RUSAGE_SELF)
-        if sys.platform == 'darwin':
-            factor = 1
-        else:
-            factor = 1024  # Linux
-        res['maxrss'] = rusage.ru_maxrss * factor / MiB
-    for key, value in res.items():
-        print(f"  {key:12.12s}: {value:10.0f} MiB")
-    return True
-
-
-# Hack: extra token to represent '{ ... }'
-CURLY_STUFF = token.N_TOKENS + 1
-token.tok_name[CURLY_STUFF] = 'CURLY_STUFF'
-
-
-def grammar_tokenizer(token_generator):
-    for tok in token_generator:
-        if tok.string == '{':
-            start = tok.start
-            nest = 1
-            accumulated = ['{']
-            for tok in token_generator:
-                accumulated.append(tok.string)
-                if tok.string == '{':
-                    nest += 1
-                elif tok.string == '}':
-                    nest -= 1
-                    if nest == 0:
-                        end = tok.end
-                        break
-            yield tokenize.TokenInfo(CURLY_STUFF, " ".join(accumulated), start, end, "")
-        else:
-            yield tok
-
-
-def simple_parser_main(parser_class):
-    argparser = argparse.ArgumentParser()
-    argparser.add_argument('-v', '--verbose', action='count', default=0,
-                           help="Print timing stats; repeat for more debug output")
-    argparser.add_argument('-q', '--quiet', action='store_true',
-                           help="Don't print the parsed program")
-    argparser.add_argument('-G', '--grammar-parser', action='store_true',
-                           help="Recognize { ... } stuff; use for meta-grammar")
-    argparser.add_argument('filename', help="Input file ('-' to use stdin)")
-
-    args = argparser.parse_args()
-    verbose = args.verbose
-    verbose_tokenizer = verbose >= 3
-    verbose_parser = verbose == 2 or verbose >= 4
-
-    t0 = time.time()
-
-    filename = args.filename
-    if filename == '' or filename == '-':
-        filename = "<stdin>"
-        file = sys.stdin
-    else:
-        file = open(args.filename)
-    try:
-        tokengen = tokenize.generate_tokens(file.readline)
-        if args.grammar_parser:
-            tokengen = grammar_tokenizer(tokengen)
-        tokenizer = Tokenizer(tokengen, verbose=verbose_tokenizer)
-        parser = parser_class(tokenizer, verbose=verbose_parser)
-        tree = parser.start()
-        try:
-            if file.isatty():
-                endpos = 0
-            else:
-                endpos = file.tell()
-        except IOError:
-            endpos = 0
-    finally:
-        if file is not sys.stdin:
-            file.close()
-
-    t1 = time.time()
-
-    if not tree:
-        err = parser.make_syntax_error(filename)
-        traceback.print_exception(err.__class__, err, None)
-        sys.exit(1)
-
-    if not args.quiet:
-        print(tree)
-
-    if verbose:
-        dt = t1 - t0
-        diag = tokenizer.diagnose()
-        nlines = diag.end[0]
-        if diag.type == token.ENDMARKER:
-            nlines -= 1
-        print(f"Total time: {dt:.3f} sec; {nlines} lines", end="")
-        if endpos:
-             print(f" ({endpos} bytes)", end="")
-        if dt:
-            print(f"; {nlines/dt:.0f} lines/sec")
-        else:
-            print()
-        print("Caches sizes:")
-        print(f"  token array : {len(tokenizer._tokens):10}")
-        print(f"  symbol cache: {len(parser._symbol_cache):10}")
-        print(f"  token cache : {len(parser._token_cache):10}")
-        print_memstats()
-
-
-argparser = argparse.ArgumentParser(prog='pegen', description="Experimental PEG-like parser generator")
-argparser.add_argument('-q', '--quiet', action='store_true', help="Don't print the parsed grammar")
-argparser.add_argument('-v', '--verbose', action='count', default=0,
-                       help="Print timing stats; repeat for more debug output")
-argparser.add_argument('-c', '--cpython', action='store_true', help="Generate C code for inclusion into CPython")
-argparser.add_argument('-o', '--output', metavar='OUT',
-                       help="Where to write the generated parser (default parse.py)")
-argparser.add_argument('filename', help="Grammar description")
-
-
-def main() -> None:
-    args = argparser.parse_args()
-    verbose = args.verbose
-    verbose_tokenizer = verbose >= 3
-    verbose_parser = verbose == 2 or verbose >= 4
-    t0 = time.time()
-
-    with open(args.filename) as file:
-        tokenizer = Tokenizer(grammar_tokenizer(tokenize.generate_tokens(file.readline)),
-                              verbose=verbose_tokenizer)
-        parser = GrammarParser(tokenizer, verbose=verbose_parser)
-        rules = parser.start()
-        if not rules:
-            err = parser.make_syntax_error(args.filename)
-            traceback.print_exception(err.__class__, err, None)
-            sys.exit(1)
-        endpos = file.tell()
-
-    if not args.quiet:
-        if args.verbose:
-            print("Raw Grammar:")
-            for rule in rules.values():
-                print(" ", repr(rule))
-        print("Clean Grammar:")
-        for rule in rules.values():
-            print(" ", rule)
-
-    output = args.output
-    if not output:
-        if args.cpython:
-            output = "parse.c"
-        else:
-            output = "parse.py"
-    with open(output, 'w') as file:
-        genr = ParserGenerator(rules, file)
-        if args.cpython:
-            genr.generate_cpython_extension(args.filename)
-        else:
-            genr.generate_python_module(args.filename)
-
-    if args.verbose:
-        print("First Graph:")
-        for src, dsts in genr.first_graph.items():
-            print(f"  {src} -> {', '.join(dsts)}")
-        print("First SCCS:")
-        for scc in genr.first_sccs:
-            print(" ", scc, end="")
-            if len(scc) > 1:
-                print("  # Indirectly left-recursive")
-            else:
-                name = next(iter(scc))
-                if name in genr.first_graph[name]:
-                    print("  # Left-recursive")
-                else:
-                    print()
-
-    t1 = time.time()
-
-    if args.verbose:
-        dt = t1 - t0
-        diag = tokenizer.diagnose()
-        nlines = diag.end[0]
-        if diag.type == token.ENDMARKER:
-            nlines -= 1
-        print(f"Total time: {dt:.3f} sec; {nlines} lines", end="")
-        if endpos:
-             print(f" ({endpos} bytes)", end="")
-        if dt:
-            print(f"; {nlines/dt:.0f} lines/sec")
-        else:
-            print()
-        print("Caches sizes:")
-        print(f"  token array : {len(tokenizer._tokens):10}")
-        print(f"  symbol cache: {len(parser._symbol_cache):10}")
-        print(f"  token cache : {len(parser._token_cache):10}")
-        if not print_memstats():
-            print("(Can't find psutil; install it for memory stats.)")
-
-
-if __name__ == '__main__':
-    main()
