+++ pegen/sccutils.py
